+++ code/controllers/GoogleAnalyticsCache_Controller.php
@@ -11,16 +11,17 @@ class GoogleAnalyticsCache_Controller extends Controller
 {
     public function index($request)
     {
+        $seconds_to_cache = Config::inst()->get('AnalyticsJS', 'cache_hours') * 60 * 60;
+
         $url = 'https://www.google-analytics.com/analytics.js';
-        $req = new RestfulService($url);
+        $req = new RestfulService($url, $seconds_to_cache);
         $conn = $req->request();
         $js = @$conn->getBody();
 
+        // if caching error, redirect to official script
         if (strlen($js) < 500) {
             return $this->Redirect('https://www.google-analytics.com/analytics.js');
         }
-
-        $seconds_to_cache = Config::inst()->get('AnalyticsJS', 'cache_hours') * 60 * 60;
 
         $this->response->addHeader('Content-type', 'application/javascript');
         $this->response->addHeader('Expires', gmdate('D, d M Y H:i:s', time() + $seconds_to_cache) . ' GMT');